airflow/requirements.txt
# airflow/requirements.txt
apache-airflow[aws]==2.6.2
opentelemetry-api==1.15.0
opentelemetry-sdk==1.15.0
opentelemetry-instrumentation==0.39b0
opentelemetry-exporter-jaeger==1.15.0
opentelemetry-instrumentation-airflow==0.37b0

